fix: keep only the pdf approach - index.py
since vercel requires .traineddata files to apply OCR to images for detect text, the image to text approach will be disregarded for now and I will leave only the pdf to text functionality, in this repository. Then I will also implement the feature of converting the image to text but with docker.